+++ CHANGELOG.md
@@ -41,6 +41,9 @@ CHANGELOG
 - `intelmq.bots.parsers.autoshun.parser`: Drop compatibility with Python 3.4.
 - `intelmq.bots.parsers.html_table.parser`: Drop compatibility with Python 3.4.
 - `intelmq.bots.parsers.shadowserver.parser`: Add suppor for MQTT feed (PR#1512).
+- `intelmq.bots.parsers.taichung.parser`:
+  - Migrate to `ParserBot`.
+  - Also parse geolocation information if available.
 
 #### Experts
 - `intelmq.bots.experts.csv_converter`: Added as converter to CSV.